Добавил prop-types
