--- src/components/WelcomeScreen/index.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-
-class WelcomeScreen extends React.PureComponent {
-
-    render() {
-        const { errorValue } = this.props;
-
-        return (
-            <>
-                <div>Добро пожаловать!!!</div>
-                <div>Сейчас у вас { errorValue } ошибки</div>
-            </>
-        )
-    }
-}
-
-export default WelcomeScreen;
